requirements.txt
Django==1.4.2
South==0.7.6
djangorestframework==2.1.3
ipdb==0.7
ipython==0.13.1
psycopg2==2.4.5
wsgiref==0.1.2
